+++ src/mark/arc.rs
@@ -9,10 +9,10 @@ use crate::visual::color::SingleColor;
 #[derive(Clone)]
 pub struct MarkArc {
     pub(crate) color: SingleColor,
-    pub(crate) opacity: f32,
+    pub(crate) opacity: f64,
     pub(crate) stroke: SingleColor,
-    pub(crate) stroke_width: f32,
-    pub(crate) inner_radius_ratio: f32,
+    pub(crate) stroke_width: f64,
+    pub(crate) inner_radius_ratio: f64,
 }
 
 impl MarkArc {
@@ -37,7 +37,7 @@ impl MarkArc {
     /// Sets the opacity of the arc mark.
     /// 
     /// Value should be between 0.0 (transparent) and 1.0 (opaque).
-    pub fn with_opacity(mut self, opacity: f32) -> Self {
+    pub fn with_opacity(mut self, opacity: f64) -> Self {
         self.opacity = opacity.clamp(0.0, 1.0);
         self
     }
@@ -49,7 +49,7 @@ impl MarkArc {
     }
 
     /// Sets the thickness of the arc's outline.
-    pub fn with_stroke_width(mut self, width: f32) -> Self {
+    pub fn with_stroke_width(mut self, width: f64) -> Self {
         self.stroke_width = width;
         self
     }
@@ -57,7 +57,7 @@ impl MarkArc {
     /// Sets the inner radius ratio (0.0 for Pie, > 0.0 for Donut).
     /// 
     /// Value is clamped between 0.0 and 1.0.
-    pub fn with_inner_radius_ratio(mut self, ratio: f32) -> Self {
+    pub fn with_inner_radius_ratio(mut self, ratio: f64) -> Self {
         self.inner_radius_ratio = ratio.clamp(0.0, 1.0);
         self
     }
